force cast to utilize index
FossilOrigin-Name: 84130ab6bfcc1a173153812e634d6df0ffbdf3c01b56113703224336d5e21874

+++ rsstg.sql
@@ -36,7 +36,7 @@ create or replace view rsstg_order as
 		from rsstg_source natural left join
 		(select source_id, count(*) as activity
 			from rsstg_post where 
-				hour = extract('hour' from now())
+				hour = extract('hour' from now())::smallint
 				and posted > now() - interval '7 days'
 			group by source_id) as act
 		where enabled
